This is not a victory
This is a surrender. None of this code works, and I don't know if it ever will.

I surrender (for now) to the complexity of the ActivityStreams spec.  It's too much for me to try to tackle right now, and it's feeling like a snipe-hunt down a rabbit-hole.

I'm checking in my current progress, in hopes that a future, braver me will pick this up and have a better plan of attack.

My best idea at this point is to give up on a strongly typed implementation, and serialize/de-serialize EVERYTHING into a map[string]interface{}.  "Types" may be "enforced" via a function library that helps the developer to parse and build otherwise infinitely flexible data structures.

May God have mercy on my code.